+++ TaskingVariables.h
@@ -1,4 +1,4 @@
- #ifndef _TASKINGVARIABLES_H  //Guard
+ /*#ifndef _TASKINGVARIABLES_H  //Guard
 #define _TASKINGVARIABLES_H
 
 struct TaskBucket
@@ -7,14 +7,14 @@ struct TaskBucket
     {
         public:
             virtual bool normal();
-    }
+    };
 
     class NavTask
     {
         public:
             virtual bool nav(geometry_msgs::Pose);
-    }
+    };
 };
 
 
-#endif
+#endif*/
