--- src/Task.cpp
@@ -1,16 +1,0 @@
-#include <iostream>
-#include "Task.h"
-#include "TaskingVariables.h"
-
-using namespace std;
-
-Task::Task()
-{
-    
-    
-}
-
-bool Task::initialize()
-{
-    return false;
-}
