--- src/Task.h
@@ -1,19 +1,0 @@
-#ifndef _TASK_H  //Guard
-#define _TASK_H
-
-using namespace std;
-
-
-class Task
-{
-    public:
-        //member functions()
-        Task();
-        virtual bool initialize();
-        TaskBucket taskBucket();
-
-    private:
-        //Data Members
-        
-};
-#endif
